+++ core/src/main/java/de/markusbordihn/adaptiveperformancetweakscore/server/ServerManager.java
@@ -24,22 +24,32 @@ import org.apache.logging.log4j.Logger;
 
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.Difficulty;
 
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.event.DifficultyChangeEvent;
 import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
 import net.minecraftforge.server.ServerLifecycleHooks;
+import net.minecraftforge.event.server.ServerAboutToStartEvent;
 import net.minecraftforge.event.server.ServerStartingEvent;
 
 import de.markusbordihn.adaptiveperformancetweakscore.Constants;
+import de.markusbordihn.adaptiveperformancetweakscore.config.CommonConfig;
 
 @EventBusSubscriber
 public class ServerManager {
 
   private static final Logger log = LogManager.getLogger(Constants.LOG_NAME);
+
+  private static final CommonConfig.Config COMMON = CommonConfig.COMMON;
+  private static double gameDifficultyFactorEasy = COMMON.gameDifficultyFactorEasy.get();
+  private static double gameDifficultyFactorNormal = COMMON.gameDifficultyFactorNormal.get();
+  private static double gameDifficultyFactorPeaceful = COMMON.gameDifficultyFactorPeaceful.get();
+  private static double gameDifficultyFactorHard = COMMON.gameDifficultyFactorHard.get();
 
   private static short ticks = 0;
   private static final short SERVER_LOAD_TICK = 1 * 20;
@@ -50,25 +60,37 @@ public class ServerManager {
   private static MinecraftServer minecraftServer = null;
   private static java.lang.Iterable<ServerLevel> serverLevels = null;
 
+  private static Difficulty gameDifficulty = Difficulty.NORMAL;
+  private static double gameDifficultyFactor = gameDifficultyFactorNormal;
+
   protected ServerManager() {
 
   }
 
   @SubscribeEvent
+  public static void handleServerAboutToStartEvent(ServerAboutToStartEvent event) {
+    gameDifficultyFactorEasy = COMMON.gameDifficultyFactorEasy.get();
+    gameDifficultyFactorNormal = COMMON.gameDifficultyFactorNormal.get();
+    gameDifficultyFactorPeaceful = COMMON.gameDifficultyFactorPeaceful.get();
+    gameDifficultyFactorHard = COMMON.gameDifficultyFactorHard.get();
+    log.info("{} Game difficult factors EASY: {}, NORMAL: {}, PEACEFUL: {} and HARD: {}",
+        Constants.LOG_PREFIX, gameDifficultyFactorEasy, gameDifficultyFactorNormal,
+        gameDifficultyFactorPeaceful, gameDifficultyFactorHard);
+  }
+
+  @SubscribeEvent
   @OnlyIn(Dist.CLIENT)
   public static void handleClientServerStartingEvent(ServerStartingEvent event) {
-    log.info("Client Server game difficulty is set to {}",
-        getMinecraftServer().getWorldData().getDifficulty());
-    log.info("Max number of local players is set to {}",
+    updateGameDifficulty(getMinecraftServer().getWorldData().getDifficulty());
+    log.info("{} Max number of local players is set to {}", Constants.LOG_PREFIX,
         getMinecraftServer().getPlayerList().getMaxPlayers());
   }
 
   @SubscribeEvent
   @OnlyIn(Dist.DEDICATED_SERVER)
   public static void handleDedicatedServerStartingEvent(ServerStartingEvent event) {
-    log.info("Dedicated Server game difficulty is set to {}",
-        getMinecraftServer().getWorldData().getDifficulty());
-    log.info("Max number of remote players is set to {}",
+    updateGameDifficulty(getMinecraftServer().getWorldData().getDifficulty());
+    log.info("{} Max number of remote players is set to {}", Constants.LOG_PREFIX,
         getMinecraftServer().getPlayerList().getMaxPlayers());
   }
 
@@ -82,6 +104,11 @@ public class ServerManager {
   @OnlyIn(Dist.DEDICATED_SERVER)
   public static void handleDedicatedServerTickEvent(TickEvent.ServerTickEvent event) {
     handleServerTickEvent(event, Dist.DEDICATED_SERVER);
+  }
+
+  @SubscribeEvent
+  public static void handleDifficultyChangeEvent(DifficultyChangeEvent event) {
+    updateGameDifficulty(event.getDifficulty());
   }
 
   public static void handleServerTickEvent(TickEvent.ServerTickEvent event, Dist dist) {
@@ -116,6 +143,39 @@ public class ServerManager {
       serverLevels = getMinecraftServer().getAllLevels();
     }
     return serverLevels;
+  }
+
+  public static Difficulty getGameDifficulty() {
+    return gameDifficulty;
+  }
+
+  public static double getGameDifficultyFactor() {
+    return gameDifficultyFactor;
+  }
+
+  private static void updateGameDifficulty(Difficulty difficulty) {
+    if (difficulty == gameDifficulty) {
+      return;
+    }
+    gameDifficulty = difficulty;
+    switch (difficulty) {
+      case EASY:
+        gameDifficultyFactor = gameDifficultyFactorEasy;
+        break;
+      case NORMAL:
+        gameDifficultyFactor = gameDifficultyFactorNormal;
+        break;
+      case PEACEFUL:
+        gameDifficultyFactor = gameDifficultyFactorPeaceful;
+        break;
+      case HARD:
+        gameDifficultyFactor = gameDifficultyFactorHard;
+        break;
+      default:
+        gameDifficultyFactor = gameDifficultyFactorNormal;
+    }
+    log.info("{} Game difficulty is set to {} with a {} factor.", Constants.LOG_PREFIX,
+        gameDifficulty, gameDifficultyFactor);
   }
 
 }